Coding

+++ src/main/java/com/vadim/sneakerstore/service/impl/FavoriteServiceImpl.java
@@ -37,7 +37,6 @@ public class FavoriteServiceImpl implements FavoriteService {
                         productId + " is not found")
         );
         return converter.convertToDto(favorite);
-     //   return null;
     }
 
     @Override
@@ -54,40 +53,37 @@ public class FavoriteServiceImpl implements FavoriteService {
         return repository.findAllByCustomerId(customerId).stream()
                 .map(converter::convertToDto)
                 .collect(Collectors.toList());
-    //    return null;
     }
 
     @Override
     @Transactional
     public FavoriteDto save(FavoriteDto favoriteDto) {
-     //   if (repository.existsByCustomerIdAndProductId(favoriteDto.getCustomerId(), favoriteDto.getProductId())) {
-     //       throw new AlreadyExistsException("Favorite with customer id = " + favoriteDto.getCustomerId()
-     //       + " and product id = " + favoriteDto.getProductId() + " already exists");
-      //  }
+        if (repository.existsByCustomerIdAndProductId(favoriteDto.getCustomerId(), favoriteDto.getProductId())) {
+            throw new AlreadyExistsException("Favorite with customer id = " + favoriteDto.getCustomerId()
+            + " and product id = " + favoriteDto.getProductId() + " already exists");
+        }
         Favorite favorite = repository.save(converter.convertToEntity(favoriteDto));
         return converter.convertToDto(favorite);
- //       return null;
     }
 
     @Override
     @Transactional
     public FavoriteDto update(FavoriteDto favoriteDto) {
-    //    if (!repository.existsByCustomerIdAndProductId(favoriteDto.getCustomerId(), favoriteDto.getProductId())) {
-   //         throw new NotFoundException("Favorite with customer id = " + favoriteDto.getCustomerId()
-    //                + " and product id = " + favoriteDto.getProductId() + " is not founa");
-    //    }
+        if (!repository.existsByCustomerIdAndProductId(favoriteDto.getCustomerId(), favoriteDto.getProductId())) {
+            throw new NotFoundException("Favorite with customer id = " + favoriteDto.getCustomerId()
+                    + " and product id = " + favoriteDto.getProductId() + " is not founa");
+        }
         Favorite favorite = repository.save(converter.convertToEntity(favoriteDto));
         return converter.convertToDto(favorite);
-       // return null;
     }
 
     @Override
     @Transactional
     public void deleteById(UUID customerId, UUID productId) {
-   //     if (!repository.existsByCustomerIdAndProductId(customerId, productId)) {
-   //         throw new NotFoundException("Favorite with customer id = " + customerId
-   //                 + " and product id = " + productId + " is not founa");
-   //     }
+        if (!repository.existsByCustomerIdAndProductId(customerId, productId)) {
+            throw new NotFoundException("Favorite with customer id = " + customerId
+                    + " and product id = " + productId + " is not founa");
+        }
         repository.deleteByCustomerIdAndProductId(customerId, productId);
     }
